requirements.txt
python-telegram-bot[httpx]==21.11.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pytest==8.3.5
watchdog==6.0.0
